fix struct deserialization
This fixes an issue where System.Text.Json can't deserialize a struct by replacing those types with our own.  This
also uncovered an issue where the deep object inspection was defaulting to compile time types intsead of runtime.
A helper was added to do the correct deep inspection and all existing calls to `BeEquivalentTo()` were audited and
updated if necessary.