fix kernel restart loop due to races
There was a race condition in the `processExited` handler on `StdioKernelTransport` that would cause the client
mapper to remove a tracked client _after_ it had already been restarted, which would then kick off a kernel restart
loop where it was constantly killing itself after it had started.  This was sometimes exacerbated by the bootstrap
API listener that was too eagerly grabbing a client.

The fix was threefold:

1. Don't fire the `processExited` event when the transport was explicitly shutdown, only when it died unexpectedly.
2. When the process _did_ exit unexpectedly, don't resolve its client just to shut it down; this could spawn extra
   processes that were never tracked.
3. Make the bootstrap API less eager and only act if a client had already been created.

Fixes #1135