Fix encoding problem (#3156)
* Fix encoding problem

StdioKernelConnector was using the Process.StandardInput StreamWriter to talk to kernels running inside the backing dotnet.exe process. The encoding for this StreamWriter however defaults to the OS default encoding for consoles. On my machine this is code page 437 - which is different from UTF8 (code page 65001). This meant that code submitted via SubmitCode commands ended up arriving misencoded in the kernels.

This was only a problem for stdin - we were correctly setting Process.StandardOutputEncoding and Process.StandardErrorEncoding to Encoding.UTF8 for the backing dotnet.exe process. But we were not setting Process.StandardInputEncoding since this property is only available in netstandard2.1 and not in nestandard2.0 (which is what we need for Microsoft.DotNet.Interactive.VisualStudio).

The fix is to use a new StreamWriter with default encoding over the same underlying Process.StandardInput.BaseStream instead. Also added a test to validate that this works correctly in both netstandard2.0 and netstandard2.1.

* Avoid disposing the StreamWriter since we can't close or dispose the stdin Stream of the process.