Fix bug related to resetting of KernelScheduler._currentlyRunningTopLevelOperation.
_currentlyRunningTopLevelOperation was not being reset in the case where a top level operation would schedule child operations. In this case, _currentlyRunningOperation was correctly being reset when each child operation was completed. However, when the parent operation was completed, we would fail to reset _currentlyRunningTopLevelOperation since we were comparing _currentlyRunningTopLevelOperation with the value of _currentlyRunningOperation (and currentlyRunningOperation would have a value null because it was reset to null when the child operation was completed above).

The above bug was breaking parenting (and command token assignment) for subsequent commands. Subsequent commands were being incorrectly parented to the stale value reflected in _currentlyRunningTopLevelOperation. And this in turn was leading to hangs.

This was an unfortunate regression triggered by a previous change that fixed a race condition. The regression happened because were were missing a test for the above scenario. I have included some tests for this as part of the current commit.