Disable shadow copy even when arcade is disabled.
Shadow copy is disabled automatically when arcade is enabled. The current change ensures that it is also disabled when arcade is disabled.

This is requierd because StdIoKernelConnector tests rely on being run from a predictable output path so that they can find the location of Microsoft.DotNet.Interactive.App.dll to launch the dotnet-interactive tool at runtime.