Configure LUT's shadow copying to include all outputs for the app project.
LUT's shadow copying only includes the ouput assembly (and no dependendent assemblies) by default for non-test projects. However, StdIoKernelConnectorTests need to find and load the app assembly from the output path (which requires all dependent assemblies present in the output folder to be copied in addition to the assembly itself).